chore: re-loacte test cases
