+++ tests/src/tests.rs
@@ -1,4 +1,4 @@
 mod cluster;
-mod mime;
 mod mutant;
 mod spore;
+mod utils;
